Final.AST

Description

This module includes functionality for creating an Abstract Syntax Tree (AST), backtraces (Backtrace), as well as helper functions for checking different aspects of the AST. The AST abstract over their kind Phase, where Phase represent the current state of the AST. For example, after parsing the AST is of Parsed Phase; after type checking with tcProgram the returned AST is of Checked Phase, indicating that the AST has been type checked.

Documentation

type Name = String

isConstructorName :: [Char] -> Bool

Check if a name is a constructor name

AST declarations

Declaration for the Abstract Syntax Tree of the language. This section contains the type, class, methods, fields and expressions represented as an AST. The AST is produced by a parser. For more information on building parsers, we recommend to read megaparsec.

data Type (p :: Phase f)

Representation of types abstracting over the Phase

Constructors

ClassType Name
Represents a class of name Name

IntType
Represents integers

BoolType
Represents booleans

Arrow
Represents a function type

Fields

tparams :: [Type p]

tresult :: Type p

UnitType
Represents the unit (void) type

newtype Program (ip :: Phase f)

The representation of a program in the form of an AST node.

Constructors

Program [ClassDef ip]
Programs are simply a list of class definitions (ClassDef) in a certain Phase

data Phase (f :: * -> *) where

Phases that have already been passed. This has been thought as going through different phases of a compiler. We assume that there is a total order between phases.

Constructors

Parsed

Fields

:: Phase Proxy

Initial status of an AST node after parsing

Checked

Fields

:: Phase Identity

Status of an AST node after type checking

data ClassDef (ip :: Phase f)

A representation of a class in the form of an AST node. As an example:

class Foo:
  val x: Int
  var y: Bool
  def main(): Int
    42

the code above, after parsing, would generate the following AST:

ClassDef {cname = "Foo"
         ,fields = [FieldDef {fname = "x"
                             ,ftype = IntType
                             ,fmod = Val}]
         ,methods = [MethodDef {mname = "main"
                               ,mparams = []
                               ,mtype = IntType
                               ,mbody = [IntLit {etype = Proxy, ival = 42}]
                               }]}

Constructors

ClassDef

Fields

cname :: Name

String that represents the name of the class

fields :: [FieldDef ip]

List of field definitions of a class

methods :: [MethodDef ip]

List of method definitions of a class

data Mod

Field qualifiers in a class. It is thought for a made up syntax such as:

class Foo:
  val x: Int
  var y: Bool

This indicates that the variable x is immutable, and y can be mutated.

Constructors

Var
Indicates that the field can be mutated

Val
Indicates that the field is immutable

data FieldDef (p :: Phase f)

Representation of a field declaration in the form of an AST node. As an example, the following code:

class Foo:
  val x: Int

could be parsed to the following field representation:

FieldDef {fname = "x"
         ,ftype = IntType
         ,fmod = Val}

Constructors

FieldDef

Fields

fname :: Name

Name of the field name

ftype :: Type p

Type of the field

fmod :: Mod

Field qualifier

isValField :: FieldDef p -> Bool

Helper function to check whether a FieldDef is immutable.

isVarField :: FieldDef p -> Bool

Helper function to check whether a FieldDef is mutable.

data Param (p :: Phase f)

Representation of parameters in the form of an AST.

Constructors

Param

Fields

pname :: Name

Name of the parameter

ptype :: Type p

Type of the parameter

data MethodDef (ip :: Phase f)

Representation of a method declaration in the form of an AST. For example:

class Foo:
  def main(): Int
    42

the code above, after parsing, would generate the following AST:

ClassDef {cname = "Foo"
         ,fields = []
         ,methods = [MethodDef {mname = "main"
                               ,mparams = []
                               ,mtype = IntType
                               ,mbody = [IntLit {etype = Proxy, ival = 42}]
                               }]}

Constructors

MethodDef

Fields

mname :: Name

Name of the method definition

mparams :: [Param ip]

List of arguments to the method

mtype :: Type ip

Return type

mbody :: Expr ip

Body of the method

commaSep :: Show t => [t] -> String

Takes a list of things that can be shown, and creates a comma separated string.

data Op

Representation of integer operations

Constructors

Add

Sub

Mul

Div

data Expr (p :: Phase f)

Representation of expressions in the form of an AST node. The language is expression-based, so there are no statements. As an example, the following identity function:

let id = \x: Int -> x
in id 42

would be parsed as this Expr:

Let {etype = Proxy
    ,name = "id"
    ,val = Lambda {etype = Proxy
                  ,params = [Param "x" IntType]
                  ,body =   FunctionCall {etype = Proxy
                                         ,target = VarAccess Proxy "id"
                                         ,args = [IntLit Proxy 42]}
                  }
    ,body  :: Expr p
    }

Constructors

BoolLit
Representation of a boolean literal

Fields

etype :: f (Type p)

Type of the expression

bval :: Bool

The Haskell Bool data constructor

IntLit
Representation of an integer literal

Fields

etype :: f (Type p)

Type of the expression

ival :: Int

Null

Fields

etype :: f (Type p)

Type of the expression

Lambda
Representation of a lambda expression

Fields

etype :: f (Type p)

Type of the expression

params :: [Param p]

List of arguments with their types (Param)

body :: Expr p

The body of the lambda abstraction

VarAccess
Representation of a variable access

Fields

etype :: f (Type p)

Type of the expression

name :: Name

Variable name

FieldAccess

Fields

etype :: f (Type p)

Type of the expression

target :: Expr p

The target in a field access, e.g., x.foo, then x is the target.

name :: Name

Variable name

Assignment

Fields

etype :: f (Type p)

Type of the expression

lhs :: Expr p

Left-hand side expression

rhs :: Expr p

Right-hand side expression

MethodCall

Fields

etype :: f (Type p)

Type of the expression

target :: Expr p

The target in a field access, e.g., x.foo, then x is the target.

name :: Name

Variable name

args :: [Expr p]

The arguments of the method call

FunctionCall

Fields

etype :: f (Type p)

Type of the expression

target :: Expr p

The target in a field access, e.g., x.foo, then x is the target.

args :: [Expr p]

The arguments of the method call

If

Fields

etype :: f (Type p)

Type of the expression

cond :: Expr p

The condition in the if-else expression

thn :: Expr p

The body of the then branch

els :: Expr p

The body of the else branch

Let

Fields

etype :: f (Type p)

Type of the expression

name :: Name

Variable name

val :: Expr p

Expression that will bound variable name with value val

body :: Expr p

The body of the lambda abstraction

BinOp

Fields

etype :: f (Type p)

Type of the expression

op :: Op

Binary operation

lhs :: Expr p

Left-hand side expression

rhs :: Expr p

Right-hand side expression

New
It is useful to decouple the type of the expression from the type of the instantiated class. This distinction becomes important whenever we have subtyping, e.g., an interface Animal where `Animal x = new Dog`

Fields

etype :: f (Type p)

Type of the expression

ty :: Type p

The class that one instantiates, e.g., `new C`

args :: [Expr p]

The arguments of the method call

Cast

Fields

etype :: f (Type p)

Type of the expression

body :: Expr p

The body of the lambda abstraction

ty :: Type p

The class that one instantiates, e.g., `new C`

Helper functions

The helper functions of this section operate on AST nodes to check for different properties. As an example, to check whether an expression is a field, instead of having to pattern match in all places, i.e.,

exampleFunction :: Expr -> Bool
exampleFunction expr =
  -- does some stuff
  ...
  case expr of
    FieldAccess expr -> True
    _                -> False

we define the isFieldAccess helper function, which checks whether a given expression is a FieldAccess:

exampleFunction :: Expr -> Bool
exampleFunction expr =
  -- does some stuff
  ...
  isFieldAccess expr

thisName :: Name

Constant for the name this, commonly used in object-oriented languages.

isArrowType :: Type p -> Bool

Checks whether a Type is a function (arrow) type

isFieldAccess :: Expr p -> Bool

Checks whether an expression is a FieldAccess.

isVarAccess :: Expr p -> Bool

Checks whether an expression is a VarAccess.

isThisAccess :: Expr p -> Bool

Checks whether an expression is a VarAccess of this.

isLVal :: Expr p -> Bool

Checks whether an expression is an lval.

isClassType :: Type p -> Bool

Helper function to check whether a Type is a class

getType :: Expr Checked -> Type Checked

Helper function to extract the type from an expression.

setType :: Type Checked -> Expr Checked -> Expr Checked

Sets the type of an expression e to t.

usesVar :: Expr p -> Name -> Bool

Helper function that checks whether an expression uses a variable. This helper function can be used to throw a warning if the variable is unused.

Backtrace declarations

newtype Backtrace

Representation of a backtrace

Constructors

Backtrace [BacktraceNode]
The representation of a backtrace consists of a list of backtrace nodes

emptyBt :: Backtrace

Helper function for starting with an empty backtrace

data BacktraceNode where

Representation of different kinds of Backtrace nodes. The backtrace nodes contain all the available information that can be shown to the developer after a crash.

Constructors

BTClass :: ClassDef p -> BacktraceNode
Represents the backtrace of a class

BTParam :: Param p -> BacktraceNode
Represents the backtrace of parameters

BTField :: FieldDef p -> BacktraceNode
Represents the backtrace of a field

BTMethod :: MethodDef p -> BacktraceNode
Represents the backtrace of a method

BTExpr :: Expr p -> BacktraceNode
Represents the backtrace of an expression

BTType :: Type p -> BacktraceNode
Represents the backtrace of a type

class Backtraceable a where

The type class defines how to create a Backtrace node, combinator backtrace, and how to push a new node to an existing backtrace.

Minimal complete definition

backtrace

Methods

backtrace :: a -> BacktraceNode

Create a backtrace node

push :: a -> Backtrace -> Backtrace

Create a new backtrace node and append it to the existing backtrace