Initial.AST

Description

This module includes functionality for creating an Abstract Syntax Tree (AST), as well as helper functions for checking different aspects of the AST.

AST declarations

Declaration for the Abstract Syntax Tree of the language. This section contains the type, class, methods, fields and expressions represented as an AST. The AST is produced by a parser. For more information on building parsers, we recommend to read megaparsec.

type Name = String

isConstructorName :: [Char] -> Bool

Check if a name is a constructor name

data Type

Representation of types

Constructors

ClassType Name

IntType

BoolType

Arrow

Fields

tparams :: [Type]

tresult :: Type

UnitType

newtype Program

The representation of a program in the form of an AST node.

Constructors

Program [ClassDef]
Programs are simply a list of class definitions (ClassDef)

data ClassDef

A representation of a class in the form of an AST node. As an example:

class Foo:
  val x: Int
  var y: Bool
  def main(): Int
    42

the code above, after parsing, would generate the following AST:

ClassDef {cname = "Foo"
         ,fields = [FieldDef {fname = "x"
                             ,ftype = IntType
                             ,fmod = Val}]
         ,methods = [MethodDef {mname = "main"
                               ,mparams = []
                               ,mtype = IntType
                               ,mbody = [IntLit {etype = Nothing, ival = 42}]
                               }]}

Constructors

ClassDef

Fields

cname :: Name

fields :: [FieldDef]

methods :: [MethodDef]

data Mod

Field qualifiers in a class. It is thought for a made up syntax such as:

class Foo:
  val x: Int
  var y: Bool

This indicates that the variable x is immutable, and y can be mutated.

Constructors

Var
Indicates that the field can be mutated

Val
Indicates that the field is immutable

data FieldDef

Representation of a field declaration in the form of an AST node. As an example, the following code:

class Foo:
  val x: Int

could be parsed to the following field representation:

FieldDef {fname = "x"
         ,ftype = IntType
         ,fmod = Val}

Constructors

FieldDef

Fields

fname :: Name

ftype :: Type

fmod :: Mod

isValField :: FieldDef -> Bool

Helper function to check whether a FieldDef is immutable.

isVarField :: FieldDef -> Bool

Helper function to check whether a FieldDef is mutable.

data Param

Representation of parameters in the form of an AST.

Constructors

Param

Fields

pname :: Name

ptype :: Type

data MethodDef

Representation of a method declaration in the form of an AST. For example:

class Foo:
  def main(): Int
    42

the code above, after parsing, would generate the following AST:

ClassDef {cname = "Foo"
         ,fields = []
         ,methods = [MethodDef {mname = "main"
                               ,mparams = []
                               ,mtype = IntType
                               ,mbody = [IntLit {etype = Nothing, ival = 42}]
                               }]}

Constructors

MethodDef

Fields

mname :: Name

mparams :: [Param]

mtype :: Type

mbody :: Expr

commaSep :: Show t => [t] -> String

Takes a list of things that can be shown, and creates a comma separated string.

data Op

Representation of integer operations

Constructors

Add

Sub

Mul

Div

data Expr

Representation of expressions in the form of an AST node. The language is expression-based, so there are no statements. As an example, the following identity function:

let id = \x: Int -> x
in id 42

generates this Expr:

Let {etype = Nothing
    ,name = "id"
    ,val = Lambda {etype = Nothing
                  ,params = [Param "x" IntType]
                  ,body =   FunctionCall {etype = Nothing
                                         ,target = VarAccess Nothing "id"
                                         ,args = [IntLit Nothing 42]}
                  }
    ,body  :: Expr p1
    }

Constructors

BoolLit
Representation of a boolean literal

Fields

etype :: Maybe Type

Type of the expression

bval :: Bool

IntLit
Representation of an integer literal

Fields

etype :: Maybe Type

Type of the expression

ival :: Int

Null
Representation of a null expression

Fields

etype :: Maybe Type

Type of the expression

Lambda
Representation of a null expression

Fields

etype :: Maybe Type

Type of the expression

params :: [Param]

List of arguments with their types (Param)

body :: Expr

The body of the lambda abstraction

VarAccess
Representation of a variable access

Fields

etype :: Maybe Type

Type of the expression

name :: Name

Variable name

FieldAccess

Fields

etype :: Maybe Type

Type of the expression

target :: Expr

The target in a field access, e.g., x.foo, then x is the target.

name :: Name

Variable name

Assignment

Fields

etype :: Maybe Type

Type of the expression

lhs :: Expr

Left-hand side expression

rhs :: Expr

Left-hand side expression

MethodCall

Fields

etype :: Maybe Type

Type of the expression

target :: Expr

The target in a field access, e.g., x.foo, then x is the target.

name :: Name

Variable name

args :: [Expr]

The arguments of the method call

FunctionCall

Fields

etype :: Maybe Type

Type of the expression

target :: Expr

The target in a field access, e.g., x.foo, then x is the target.

args :: [Expr]

The arguments of the method call

If

Fields

etype :: Maybe Type

Type of the expression

cond :: Expr

The condition in the if-else expression

thn :: Expr

The body of the then branch

els :: Expr

The body of the else branch

Let

Fields

etype :: Maybe Type

Type of the expression

name :: Name

Variable name

val :: Expr

Expression that will bound variable name with value val

body :: Expr

The body of the lambda abstraction

BinOp

Fields

etype :: Maybe Type

Type of the expression

op :: Op

Binary operation

lhs :: Expr

Left-hand side expression

rhs :: Expr

Left-hand side expression

New
It is useful to decouple the type of the expression from the type of the instantiated class. This distinction becomes important whenever we have subtyping, e.g., an interface Animal where `Animal x = new Dog`

Fields

etype :: Maybe Type

Type of the expression

ty :: Type

The class that one instantiates, e.g., `new C`

args :: [Expr]

The arguments of the method call

Cast

Fields

etype :: Maybe Type

Type of the expression

body :: Expr

The body of the lambda abstraction

ty :: Type

The class that one instantiates, e.g., `new C`

Helper functions

The helper functions of this section operate on AST nodes to check for different properties. As an example, to check whether an expression is a field, instead of having to pattern match in all places, i.e.,

exampleFunction :: Expr -> Bool
exampleFunction expr =
  -- does some stuff
  ...
  case expr of
    FieldAccess expr -> True
    _                -> False

we define the isFieldAccess helper function, which checks whether a given expression is a FieldAccess:

exampleFunction :: Expr -> Bool
exampleFunction expr =
  -- does some stuff
  ...
  isFieldAccess expr

thisName :: Name

Constant for the name this, commonly used in object-oriented languages.

isArrowType :: Type -> Bool

Checks whether a Type is a function (arrow) type

isFieldAccess :: Expr -> Bool

Checks whether an expression is a FieldAccess.

isVarAccess :: Expr -> Bool

Checks whether an expression is a VarAccess.

isThisAccess :: Expr -> Bool

Checks whether an expression is a VarAccess of this.

isLVal :: Expr -> Bool

Checks whether an expression is an lval.

isClassType :: Type -> Bool

Helper function to check whether a Type is a class

getType :: Expr -> Type

Helper function to extract the type from an expression.

setType :: Type -> Expr -> Expr

Sets the type of an expression e to t.