Initial.Typechecker

Description

This module includes everything you need to get started type checking a program. To build the Abstract Syntax Tree (AST), please import and build the AST from Initial.AST.

The main entry point to the type checker is the combinator tcProgram, which takes an AST and returns either an error, or the typed program. For example, for the following program (using a made up syntax):

class C
  val f: Foo

should be parsed to generate this AST:

testClass1 =
 ClassDef {cname = "C"
          ,fields = [FieldDef {fmod = Val, fname = "f", ftype = ClassType "Foo"}]
          ,methods = []}

To type check the AST, run the tcProgram combinator as follows:

tcProgram testClass1

Documentation

data TCError

Data declaration of available errors. Value constructors are used to create statically known errors. For example:

UnknownClassError  (Name c)

creates a UnknownClassError. This error should be created whenever there is a class whose declaration is unknown or inexistent.

Constructors

UnknownClassError Name
Reference of a class that does not exists

UnknownFieldError Name
Reference of a field that does not exists

UnknownMethodError Name
Reference of a method that does not exists

UnboundVariableError Name
Unbound variable

TypeMismatchError Type Type
Type mismatch error, the first Type refers to the formal type argument, the second Type refers to the actual type argument.

ImmutableFieldError Expr
Immutable field error, used when someone violates immutability

NonLValError Expr
Error to indicate that a one cannot assign a value to expression Expr

PrimitiveNullError Type
Error indicating that the return type cannot be Null

NonClassTypeError Type
Used to indicate that Type is not of a class type

NonArrowTypeError Type
Expecting a function (arrow) type but got another type instead.

ConstructorCallError Type
Tried to call a constructor outside of instantiation

UninferrableError Expr
Cannot infer type of Expr

data Env

Environment. The Env is used during type checking, and is updated as the type checker runs. Most likely, one uses the Reader monad to hide details of how the environment is updated, via the common local function.

Constructors

Env

Fields

ctable :: Map Name ClassDef

vartable :: Map Name Type

constructor :: Bool

emptyEnv :: Env

Generates an empty environment.

lookupClass :: Env -> Name -> Except TCError ClassDef

Helper function to lookup a class given a Name and an Env. For example:

typecheck env (ClassType c) = do
  _ <- lookupClass env c
  return $ ClassType c

lookupField :: Env -> Type -> Name -> Except TCError FieldDef

Find a field declaration by its Type (ty) and field name f

lookupMethod :: Env -> Type -> Name -> Except TCError MethodDef

Find a method declaration by its Type (ty) and field name f

lookupVar :: Env -> Name -> Except TCError Type

Look up a variable by its Name in the Env, returning an exception with the type checking error, TCError, or the Type of the variable x.

genEnv :: Program -> Env

Generates an environment (symbol's table) from a Program,

addVariable :: Env -> Name -> Type -> Env

Add a variable name and its type to the environment Env.

addParameters :: Env -> [Param] -> Env

Add a list of parameters, Param, to the environment.

tcProgram :: Program -> Either TCError Program

Main entry point of the type checker. This function type checks an AST returning either an error or a well-typed program. For instance, assuming the following made up language: > > class C > val f: Foo >

it should be parsed to generate the following AST:

testClass1 =
 ClassDef {cname = "C"
          ,fields = [FieldDef {fmod = Val, fname = "f", ftype = ClassType "Foo"}]
          ,methods = []}

To type check the AST, run the tcProgram combinator as follows:

tcProgram testClass1

which either returns an error or the resulting typed AST.

class Typecheckable a where

The type class defines how to type check an AST node.

Methods

typecheck :: Env -> a -> Except TCError a

Type check the well-formedness of an AST node.

hasType :: Env -> Expr -> Type -> Except TCError Expr

This combinator is used whenever a certain type is expected. This function is quite important. Here follows an example:

doTypecheck mdef@(MethodDef {mparams, mbody, mtype}) = do
  -- typecheck the well-formedness of types of method parameters
  mparams' <- mapM typecheck mparams
  mtype' <- typecheck mtype

  -- extend environment with method parameters and typecheck body
  mbody' <- local (addParameters mparams) $ hasType mbody mtype'
  ...

in the last line, because we are type checking a method declaration, it is statically known what should be the return type of the function body. In these cases, one should use the hasType combinator.

testClass1 :: ClassDef

Test programs of a class with a single field. This program is the AST equivalent of the following syntax:

class C
  val f: Foo

testClass2 :: ClassDef

Test program with a class, field, method, and variable access. The class Bar does not exist in the environment. The variable access is unbound.

This program is the AST equivalent of the following syntax:

class D
  val g: Bar
  def m(): Int
    x

testClass3 :: [ClassDef]

Test program with a two classes, field, method, and variable access. The class declaration are duplicated.

This program is the AST equivalent of the following syntax:

class D
  val g: Bar
  def m(): Int
    x

class D
  val g: Bar
  def m(): Int
    x

testProgram :: Program

testValidProgram :: Program

testSuite :: IO ()

Test suite that runs testProgram.