MultiError.Typechecker

Description

This module includes everything you need to get started type checking a program. To build the Abstract Syntax Tree (AST), please import and build the AST from MultiError.AST.

The main entry point to the type checker is the combinator tcProgram, which takes an AST and returns either a list of errors, or the typed program. For example, for the following program (using a made up syntax):

class C
  val f: Foo

should be parsed to generate this AST:

testClass1 =
 ClassDef {cname = "C"
          ,fields = [FieldDef {fmod = Val, fname = "f", ftype = ClassType "Foo"}]
          ,methods = []}

To type check the AST, run the tcProgram combinator as follows:

tcProgram testClass1

This is an increment on top of the Typechecker module, that refactors the type checker to be able to throw multiple errors.

Documentation

type TypecheckM a = forall m. (MonadReader Env m, MonadError TCErrors m) => m a

The type checking monad. The type checking monad is the stacking of the Reader and Exception monads.

(<:>) :: Semigroup e => Either e a -> Either e b -> Either e (a, b)

The function <:> takes two Either monads and returns an error if one of them is an error or aggregates both results. For example:

let error = Left "Error" <:> Right 42
let errors = Left "Error" <:> Left "Error2"
let valid = Right "42" <:> Right "0"

evaluates error = Left Error, errors = Left ErrorError2, and valid = 420.

(<&>) :: (Semigroup e, MonadError e m) => m a -> m b -> m (a, b)

Forks two computations in the Except monad, and either returns both of their results, or aggregates the errors of one or both of the computations. For example:

    (fields', methods') <- forkM precheck fields <&>
                           forkM precheck methods

In this example, if the evaluation of forkM precheck fields and and forkM precheck methods return errors, we aggregate them using <:>. If only one of them fails, then return the single error. If both computation succeed, return a monad wrapped around the tuple with both results.

forkM :: (Semigroup e, MonadError e m) => (a -> m b) -> [a] -> m [b]

Allows typechecking a list of items, collecting error messages from all of them.

newtype TCErrors

Declaration of type checking errors. An error will (usually) be created using the helper function tcError. As an example:

tcError $ DuplicateClassError (Name "Foo")

throws an error that indicates that the class is defined multiple times.

Constructors

TCErrors (NonEmpty TCError)

data TCError

Declaration of a type checking error, where Error represents the current type checking error and Backtrace the up-to-date backtrace.

Constructors

TCError
Type checking error value constructor

Fields

Error

Current type checking error

Backtrace

Backtrace of the type checker

tcError :: Error -> TypecheckM a

Throw a type checking Error

data Error

Data declaration of available errors. Value constructors are used to create statically known errors. For example:

UnknownClassError  (Name c)

creates a UnknownClassError. This error should be created whenever there is a class whose declaration is unknown or inexistent.

Constructors

UnknownClassError Name
Reference of a class that does not exists

UnknownFieldError Name
Reference of a field that does not exists

UnknownMethodError Name
Reference of a method that does not exists

UnboundVariableError Name
Unbound variable

TypeMismatchError Type Type
Type mismatch error, the first Type refers to the formal type argument, the second Type refers to the actual type argument.

ImmutableFieldError Expr
Immutable field error, used when someone violates immutability

NonLValError Expr
Error to indicate that a one cannot assign a value to expression Expr

PrimitiveNullError Type
Error indicating that the return type cannot be Null

NonClassTypeError Type
Used to indicate that Type is not of a class type

NonArrowTypeError Type
Expecting a function (arrow) type but got another type instead.

ConstructorCallError Type
Tried to call a constructor outside of instantiation

UninferrableError Expr
Cannot infer type of Expr

data Env

Environment. The Env is used during type checking, and is updated as the type checker runs. Most likely, one uses the Reader monad to hide details of how the environment is updated, via the common local function.

Constructors

Env

Fields

ctable :: Map Name ClassDef

vartable :: Map Name Type

bt :: Backtrace

constructor :: Bool

setConstructor :: Name -> Env -> Env

Conditionally update the environment to track if we are in a constructor method.

emptyEnv :: Env

Generates an empty environment.

lookupClass :: Name -> Env -> Maybe ClassDef

Helper function to lookup a class given a Name and an Env. Usually it relies on the Reader monad, so that passing the Env can be omitted. For example:

findClass :: Type -> TypecheckM ClassDef
findClass ty@(ClassType c) = do
  cls <- asks $ lookupClass c
  case cls of
    Just cdef -> return cdef
    Nothing -> tcError $ UnknownClassError c
findClass ty = tcError $ NonClassTypeError ty

In this function (findClass), the Reader function asks injects the Reader monad as the last argument. More details in the paper.

lookupVar :: Name -> Env -> Maybe Type

Look up a variable by its Name in the Env, returning an option type indicating whether the variable was found or not.

findClass :: Type -> TypecheckM ClassDef

Find a class declaration by its Type

findMethod :: Type -> Name -> TypecheckM MethodDef

Find a method declaration by its Type and method name m

findField :: Type -> Name -> TypecheckM FieldDef

Find a field declaration by its Type (ty) and field name f

findVar :: Name -> TypecheckM Type

Find a variable in the environment by its name x

genEnv :: Program -> Env

Generates an environment (symbol's table) from a Program,

addVariable :: Name -> Type -> Env -> Env

Add a variable name and its type to the environment Env.

addParameters :: [Param] -> Env -> Env

Add a list of parameters, Param, to the environment.

tcProgram :: Program -> Either TCErrors Program

Main entry point of the type checker. This function type checks an AST returning either a list of errors or a well-typed program. For instance, assuming the following made up language: > > class C > val f: Foo >

it should be parsed to generate the following AST:

testClass1 =
 ClassDef {cname = "C"
          ,fields = [FieldDef {fmod = Val, fname = "f", ftype = ClassType "Foo"}]
          ,methods = []}

To type check the AST, run the tcProgram combinator as follows:

tcProgram testClass1

which either returns a list of errors or the resulting typed AST.

class Typecheckable a where

The type class defines how to type check an AST node.

Minimal complete definition

doTypecheck

Methods

doTypecheck :: a -> TypecheckM a

Type check the well-formedness of an AST node.

typecheck :: Backtraceable a => a -> TypecheckM a

Type check an AST node, updating the environment's backtrace.

hasType :: Expr -> Type -> TypecheckM Expr

This combinator is used whenever a certain type is expected. This function is quite important. Here follows an example:

doTypecheck mdef@(MethodDef {mparams, mbody, mtype}) = do
  -- typecheck the well-formedness of types of method parameters
  mparams' <- mapM typecheck mparams
  mtype' <- typecheck mtype

  -- extend environment with method parameters and typecheck body
  mbody' <- local (addParameters mparams) $ hasType mbody mtype'
  ...

in the last line, because we are type checking a method declaration, it is statically known what should be the return type of the function body. In these cases, one should use the hasType combinator.

testClass1 :: ClassDef

Class definition for didactic purposes. This AST represents the following class, which is named C, contains an immutable field f of type Foo:

class C:
  val f: Foo

This class is ill-typed, as there is no declaration of Foo anywhere. To check how to type checker catches this error, run:

tcProgram (Program [testClass1])

testClass2 :: ClassDef

Test program with a class, field, method, and variable access. The class Bar does not exist in the environment. The variable access is unbound.

This program is the AST equivalent of the following syntax:

class D
  val g: Bar
  def m(): Int
    x

testProgram :: Program

testSuite :: IO ()

Test suite that runs testProgram.